moved the very general batch calls outside of functions.py
